Uploading semi-complete PS3
Uploading the work I've done so far for PS3. The only parts left to do is re-run the cprofile and then implement a data structure with better asymptotic run time than what is currently implemented.

In this upload have edited the sweep line algorithm so that instead of keeping all horizontal wires in the range_index throughout the sweep they are only kept in the range index between their start and their end point.