Completed AVL tree implementation of PS3
Have uploaded an AVL tree implementation to solve PS3.

Read in lecture notes from a Cambridge course - that I will work through after this course - that it is a convention to name all functions that are only called from within the class starting with an underscore. AS such have tried to stick to this convention throughout the class.

Took longer than anticipated as originally had developed an AVL Tree that changed RangeIndex too much. This meant that the JSONP trace kept crashing as I'd changed the number of arguments inputted to some of the functions.

Most of the work done is contained within the AVL Node class. This was the fix I came up with to leave RangeIndex as unchanged as possible. The way I implemented it left the number of arguments unchanged for all the functions in Range index and only meant a slight adjustment to trace range index.